pyproject.toml
[project]
name = "zplatipld"
version = "0.1.0"
description = ""
authors = [{ name = "Suender Oliveira", email = "suender@ibm.com" }]
requires-python = "~=3.11"
readme = "README.md"
dependencies = [
    "flask[async]>=3.1.0,<4",
    "flask-socketio>=5.5.1,<6",
    "flask-login>=0.6.3,<0.7",
    "jinja2>=3.1.5,<4",
    "python-dotenv>=1.0.1,<2",
    "schedule>=1.2.2,<2",
    "sqlalchemy>=2.0.37,<3",
    "asyncssh>=2.19.0,<3",
    "pandas>=2.2.3,<3",
    "flask-wtf>=1.2.2,<2",
    "requests>=2.32.3,<3",
    "pydantic-settings>=2.10.0",
    "aiofiles>=24.1.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
